application layer implemented database connecton
